Refactored ApiCalls class to return AxiosResponses
Refactored API call methods to return AxiosResponses as this allows the caller of the method to access HTTP status codes. This means that error handling will be done by the caller instead of the utility class. I think this is the better choice as two different callers probably want to handle errors in their own way. The request code, which is always the same, can still be reused.