Removed <form> from login/page.tsx as it is not used
The <form> introduces logic behind the scene related to performing a post request, which leads to extra overhead. This is unnecessary as we track the fields values in React states. We then use these states when performing the request. Therefore we completely bypass the default behaviour of the <form> - basically rendering it useless for more than semantic purposes.